Feature 0 : Ingestion sécurisée de Film, Collection, Genre, FilmGenres

+++ data_tools/0_shared/tables/genres.py
@@ -6,7 +6,7 @@ Note : le CSV genres.csv ne contient pas id_genre — il sera généré
        automatiquement à l'ingestion (range 1-19, cohérent avec film_genres.csv).
 """
 
-from sqlalchemy import Column, SmallInteger, String
+from sqlalchemy import Column, Integer, String
 from sqlalchemy.orm import relationship
 from tables.base import Base
 
@@ -21,8 +21,8 @@ class Genre(Base):
 
     __tablename__ = "genres"
 
-    # Primary key using a small integer to save disk space and memory
-    id_genre = Column(SmallInteger, primary_key=True, autoincrement=True)
+    # Primary key
+    id_genre = Column(Integer, primary_key=True, autoincrement=True)
 
     # Genre name restricted to 50 characters, must be unique
     genre_name = Column(String(50), nullable=False, unique=True)
